Separate nonstandard code tests and make the TTC output a fixture
