+++ packages/augmentation-lambda/tests/conftest.py
@@ -26,28 +26,32 @@ TEST_EICR_PATH = (
     / "basic_test_eicr.xml"
 )
 
-TEST_TTC_OUTPUT = {
-    "persistence_id": TEST_PERSISTENCE_ID,
-    "eicr_metadata": {},
-    "schematron_errors": {
-        "Lab Test Name Resulted": [
-            {
-                "schematron_error": "Text to Code: Lab Test Name Resulted does not have a @code attribute",
-                "schematron_error_xpath": "/ClinicalDocument/component/structuredBody/component/section/entry/component/observation",
-                "field_type": "Lab Test Name Resulted",
-                "new_translation": {
-                    "code": "109224-6",
-                    "code_system": "2.16.840.1.113883.6.1",
-                    "code_system_name": "LOINC",
-                    "display_name": "Weed Allergen Mix 3 IgE Ab",
-                    "value_set": None,
-                    "value_set_version": None,
-                    "original_text": "A custom code in original text.",
-                },
-            }
-        ]
-    },
-}
+
+@pytest.fixture
+def test_ttc_output() -> dict:
+    """A test example of the output of the TTC lambda."""
+    return {
+        "persistence_id": TEST_PERSISTENCE_ID,
+        "eicr_metadata": {},
+        "schematron_errors": {
+            "Lab Test Name Resulted": [
+                {
+                    "schematron_error": "Text to Code: Lab Test Name Resulted does not have a @code attribute",
+                    "schematron_error_xpath": "/ClinicalDocument/component/structuredBody/component/section/entry/component/observation",
+                    "field_type": "Lab Test Name Resulted",
+                    "new_translation": {
+                        "code": "109224-6",
+                        "code_system": "2.16.840.1.113883.6.1",
+                        "code_system_name": "LOINC",
+                        "display_name": "Weed Allergen Mix 3 IgE Ab",
+                        "value_set": None,
+                        "value_set_version": None,
+                        "original_text": "A custom code in original text.",
+                    },
+                }
+            ]
+        },
+    }
 
 
 def pytest_configure() -> None:
@@ -90,7 +94,7 @@ def example_s3_event_payload() -> dict:
     }
 
 
-@pytest.fixture
+@pytest.fixture(scope="function")
 def example_sqs_event(example_s3_event_payload: dict) -> dict:
     """Full SQS event that mimics real Lambda input."""
     return {
@@ -116,7 +120,7 @@ def example_sqs_event(example_s3_event_payload: dict) -> dict:
 
 
 @pytest.fixture(scope="function")
-def mock_aws_setup(monkeypatch: pytest.MonkeyPatch) -> boto3.client:
+def mock_aws_setup(monkeypatch: pytest.MonkeyPatch, test_ttc_output: dict) -> boto3.client:
     """Setup test AWS environment with moto mock S3."""
     with moto.mock_aws():
         monkeypatch.setenv("AWS_REGION", AWS_REGION)
@@ -147,7 +151,7 @@ def mock_aws_setup(monkeypatch: pytest.MonkeyPatch) -> boto3.client:
         s3.put_object(
             Bucket=S3_BUCKET,
             Key=f"{TTC_OUTPUT_PREFIX}{TEST_PERSISTENCE_ID}",
-            Body=json.dumps(TEST_TTC_OUTPUT),
+            Body=json.dumps(test_ttc_output),
         )
 
         # Reset cached S3 client so Lambda creates a new one inside moto context
